Fix update_network_statistics() and the network displays.
They were all totally wrong.  I was using microseconds instead of
milliseconds, I was using much too large a moving average window,
and I was displaying things stupidly.  This is the first version
that actually is USEFUL as a HUD in its intended role: tell me
WTF is going on at a glance.