Add -X because the speed of my laptop's egress is not the speed of the house DSL
Finally realized that querying the media flags to automatically guess
the speed is cute and all but it is frequently irrelevant.  Like, in
my life for instance.  Our crapoid TelMex DSL can maybe do 1.5mbit/sec
on a good day with a strong tail wind, but hardly ever even that.  The
nic in this laptop is 100mbit so if i rely on that for my max speed to
calculate percentages they are almost always zero.  Only when I do
some big file xfer over the LAN do I actually see anything on the
network meter.

Duh.